refactor: simplify player init
